Refactor packages
Moved and renamed some packages
