+++ src/main/java/ru/kingsbot/Bot.java
@@ -3,14 +3,16 @@ package ru.kingsbot;
 import com.google.gson.Gson;
 import lombok.Getter;
 import lombok.extern.log4j.Log4j2;
-import org.apache.logging.log4j.core.config.Property;
-import ru.kingsbot.api.LongPool;
-import ru.kingsbot.api.TransportClient;
-import ru.kingsbot.api.keyboard.Keyboards;
-import ru.kingsbot.attack.BossAttackMap;
-import ru.kingsbot.boss.BossMap;
+import ru.kingsbot.client.LongPool;
+import ru.kingsbot.client.TransportClient;
+import ru.kingsbot.command.keyboard.Keyboards;
+import ru.kingsbot.game.attack.BossAttackMap;
+import ru.kingsbot.game.boss.BossMap;
 import ru.kingsbot.command.CommandMap;
-import ru.kingsbot.service.*;
+import ru.kingsbot.service.ConversationService;
+import ru.kingsbot.service.DonateService;
+import ru.kingsbot.service.MarketService;
+import ru.kingsbot.service.PlayerService;
 import ru.kingsbot.utils.HibernateUtil;
 import ru.kingsbot.utils.Utils;
