+++ src/main/java/ru/kingsbot/client/TransportClient.java
@@ -1,4 +1,4 @@
-package ru.kingsbot.api;
+package ru.kingsbot.client;
 
 import lombok.extern.log4j.Log4j2;
 import org.apache.http.HttpResponse;
