+++ src/main/java/ru/kingsbot/command/chat/InfoCommand.java
@@ -1,7 +1,7 @@
 package ru.kingsbot.command.chat;
 
 import ru.kingsbot.Emoji;
-import ru.kingsbot.api.keyboard.Keyboards;
+import ru.kingsbot.command.keyboard.Keyboards;
 import ru.kingsbot.command.Command;
 import ru.kingsbot.entity.Player;
 import ru.kingsbot.entity.building.Capitol;
