+++ src/main/java/ru/kingsbot/command/group/clan/ClanRequestCommand.java
@@ -1,7 +1,7 @@
 package ru.kingsbot.command.group.clan;
 
 import ru.kingsbot.Emoji;
-import ru.kingsbot.api.keyboard.Keyboards;
+import ru.kingsbot.command.keyboard.Keyboards;
 import ru.kingsbot.command.Command;
 import ru.kingsbot.entity.Player;
 import ru.kingsbot.entity.clan.Clan;
