+++ src/main/java/ru/kingsbot/command/keyboard/Color.java
@@ -1,4 +1,4 @@
-package ru.kingsbot.api.keyboard;
+package ru.kingsbot.command.keyboard;
 
 public enum Color {
