+++ src/main/java/ru/kingsbot/command/keyboard/Keyboards.java
@@ -1,4 +1,4 @@
-package ru.kingsbot.api.keyboard;
+package ru.kingsbot.command.keyboard;
 
 import lombok.AccessLevel;
 import lombok.Getter;
