+++ src/main/java/ru/kingsbot/game/Tutorial.java
@@ -1,9 +1,9 @@
-package ru.kingsbot.tutorial;
+package ru.kingsbot.game;
 
 import ru.kingsbot.Bot;
-import ru.kingsbot.api.keyboard.Button;
-import ru.kingsbot.api.keyboard.Color;
-import ru.kingsbot.api.keyboard.Keyboard;
+import ru.kingsbot.command.keyboard.Button;
+import ru.kingsbot.command.keyboard.Color;
+import ru.kingsbot.command.keyboard.Keyboard;
 import ru.kingsbot.entity.Player;
 import ru.kingsbot.service.PlayerService;
