+++ src/main/java/ru/kingsbot/game/attack/BossAttack.java
