+++ src/main/java/ru/kingsbot/game/attack/BossAttackMap.java
