+++ src/main/java/ru/kingsbot/game/boss/Arlong.java
