+++ src/main/java/ru/kingsbot/game/boss/Boss.java
