+++ src/main/java/ru/kingsbot/game/boss/BossMap.java
