+++ src/main/java/ru/kingsbot/game/boss/Crocodile.java
