+++ src/main/java/ru/kingsbot/game/boss/Krieg.java
