+++ src/main/java/ru/kingsbot/service/PlayerService.java
@@ -5,9 +5,9 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import lombok.extern.log4j.Log4j2;
-import ru.kingsbot.api.ApiRequest;
-import ru.kingsbot.api.TransportClient;
-import ru.kingsbot.api.keyboard.Keyboard;
+import ru.kingsbot.client.ApiRequest;
+import ru.kingsbot.client.TransportClient;
+import ru.kingsbot.command.keyboard.Keyboard;
 import ru.kingsbot.entity.Player;
 import ru.kingsbot.entity.clan.Clan;
 import ru.kingsbot.repository.PlayerRepository;
